+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Helper for drawing rectangles in matplotlib
 - Multiclass AUC calculation from [Hand and Till, 2001], [Provost and Domingos, 2000]
 - Multicolumn label encoder helper
-- Utility to add MCAR missing values to a data matrix
+- Utility to add MCAR, MAR and NMAR missing values to a data matrix
+- Simple kNN implementation which is robust against missing values
 
 ### Updated
 - Confusion matrix plotting utility to work on axes objects and to be more